demo2/requirements.txt
Flask==1.1.2
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==20.0.4
google-cloud-vision==1.0.0
google-cloud-storage==1.30.0
google-cloud-pubsub==1.7.0
Wand==0.6.2
pillow